♻️ refactor: move PublishMessage handler to exchanges.go and remove messages.go
